+++ src/BookStore.Domain/Interfaces/IBookService.cs
@@ -10,11 +10,11 @@ namespace BookStore.Domain.Interfaces
     public interface IBookService: IDisposable //IDisposable: utilizzato per il rilascio della memoria
     {
         Task<IEnumerable<Book>> GetAll();
-        Task<Book> GetById(int id);
+        Task<Book> GetById(long id);
         Task<Book> Add(Book book);
         Task<Book> Update(Book book);
         Task<Book> Remove(Book book);
-        Task<IEnumerable<Book>> GetBooksByCategory(int categoryId);
+        Task<IEnumerable<Book>> GetBooksByCategory(long categoryId);
         Task<IEnumerable<Book>> Search(string bookName);
         Task<IEnumerable<Book>> SearchBookWithCategory(string searchValue);
     }
